# lib/api-zod/src/generated/types/projectMediaInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sema audio-description review API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectMediaInput {
  /** @minLength 1 */
  objectPath: string;
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
  /** @minimum 0 */
  duration?: number;
}