# lib/api-zod/src/generated/types/reviewWorkspace.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sema audio-description review API
 * OpenAPI spec version: 0.1.0
 */
import type { DescriptionCandidate } from './descriptionCandidate';
import type { NarrationWindow } from './narrationWindow';
import type { ProcessStatus } from './processStatus';
import type { Project } from './project';
import type { ReviewDecision } from './reviewDecision';
import type { VisualBeat } from './visualBeat';

export interface ReviewWorkspace {
  project: Project;
  beats: VisualBeat[];
  windows: NarrationWindow[];
  candidates: DescriptionCandidate[];
  decisions: ReviewDecision[];
  process: ProcessStatus;
}